requirements.txt
appdirs==1.4.3
argh==0.26.2
atomicwrites==1.3.0
attrs==19.1.0
black==19.3b0
Click==7.0
colorama==0.4.1
Django==2.2.1
djangorestframework==3.9.3
docopt==0.6.2
gunicorn==19.9.0
more-itertools==7.0.0
mypy==0.701
mypy-extensions==0.4.1
pathtools==0.1.2
pluggy==0.9.0
psycopg2==2.8.2
py==1.8.0
pytest==4.4.1
pytest-watch==4.2.0
pytz==2019.1
PyYAML==5.1
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.3.4
uWSGI==2.0.18
watchdog==0.9.0
